The Book of Mirdad quote
Love is the law of God. You live that you may learn to love. You love that you may learn to live. No other lesson is required of Man. And what os ot to love but for the lover to absorb forever the beloved so that the twain be one? And whom, or what, is one to love? Is one to choose a certain leaf upon the Tree of Life and pour upon it all one`s heart? What of the branch that bears the leaf? What of the stem that holds the branch? What of the bark that shields the stem? What of the roots that feed the bark, the stem, the branches and the leaves? What of the soil embosoming the roots? What of the Sun,and sea, and air that fertilize the soil? If one small leaf upon a tree be worthy of your love how much more so the entire tree in its entirity? The love that singles out a fraction of the whole foredooms itself to grief. You say But there be leaves and leaves upon a single tree. Some are healthy, some are sick; some are beautiful, some, ugly; some are giants , some are dwarfs. How can we help but pick and choose. I say to you, Out of the paleness of the sick proceeds the freshness of the healthy. I further say to you that ugliness is Beauty`s palette, paint and brush; and that the dwarf would not have been a dwarf had he not given of his stature to the giant. You are the tree of Life. Beware of fractioning yourselves. Set not a fruit against a fruit, a leaf against a leaf, a bough against a bough; nor set the stem against the roots; nor set the tree against the mother- soil. That is precisely what you do when you love one part more than the rest, or to the exclusion of the rest. You are the Tree of Life. Your roots are everywhere. Your boughs and leaves are everywhere. Your fruits are in every mouth. Whatever be the fruits upon that tree; whatever be its boughs and leaves; whatever be the roots; they are your fruits; they are your leaves and boughs; they are your roots. If you would have the tree bear sweet and fragrant fruit, if you would have it ever strong and green, see to the sap wherewith you feed the roots. Love is the Sap of Life. While Hatred is the pus of Death. But Love, like blood, must circulate unhindered in the veins. Repress the blood, and it becomes a menace and a plague. And what is Hate but Love repressed, or Love withheld, therefore becoming such a deadly poison both to the feeder and the fed; both to the hater and to that he hates. A yellow leaf upon your tree of life is but a Love-weaned leaf, Blame not the yellow leaf. A withered bough is but a Love-starved bough. Blame not the withered bough. A putrid fruit is but a Hatred-suckled fruit. Blame not the putrid fruit. But rather blame your blind and stingy heart that would dole out the sap of life to few and would deny it to many, thereby denying it to itself. No love is possible except by the love of self. No self is real save the All-embracing Self. Therefore is God all Love, because he loves himself. So long as you are pained by Love, you have not found your real self, nor have you found the golden key of Love. Because you love an ephemeral self, your love is ephemeral. The love of a man for woman is not love. It is thereof a very distant token. The love of a parent for the child is but the threshold to Love`s holy temple. Till every man be every woman`s lover, and the reverse: till every child be every parent`s child, and the reverse, let men and women brag of flesh and bone clinging to flesh and bone, but never speak the sacred name of Love.For that is blasphemy. You have no friends so long as you can count a single man a foe. The heart that harbours enmity how can it be a safe abode for friendship? You do noy know the Joy of Love so long as there is hatred in your hearts. Were you to feed all things the sap of Life except a certain tiny worm, that certain tin worm alone would embitter your life. For in loving anything, or anyone, you love in truth but yourselves. Likewise in hating anything, or anyone, you hate in truth but yourselves. For that which you hate is bound up inseparably with that which you love, like the face and the reverse side of the same coin. If you would be honest with yourselves, then must you love what you hate and what hates you before you love what you love and what loves you. Love is not a virtue. Love is a neccessity; more so than bread and water; more so than light and air. Let no one pride himself on loving. But rather breathe in Love and breathe it out just as unconsciously and freely as you breathe in the air and breathe it out. For Love needs no one to exalt it. Love will exalt the heart that it finds worthy of itself. Seek no rewards for Love. Love is reward sufficient unto LOve, as Hate is punishment sufficient unto Hate. Nor keep any accounts with Love. For Love accounts to no one but itself. Love neither lends nor borrows; Love neither buys nor sells; but when it gives, it gives its all; and when it takes, it takes its all. Its very taking is a giving. Its very giving is a taking. Therefore is it the same to-day, to-morrow and forevermore. In a short while there will be more to come.